refactor progress HoCing
